docs: modify documentation and examples
Update documentations and message descriptions
Add and modify examples for R9710 and R9711 to better reflect the errors
Change the adding of function name to the error message inside the
checker
