+++ examples/custom_checkers/r9710_inconsistent_returns.py
@@ -1,5 +1,5 @@
 import math
-from typing import List, Optional
+from typing import Optional
 
 
 def add_sqrts(x: float, y: float) -> Optional[float]:
@@ -11,30 +11,14 @@ def add_sqrts(x: float, y: float) -> Optional[float]:
         return  # Error: this should be `return None` instead.
 
 
-def index_of(numbers: List[int], n: int) -> Optional[int]:
-    """Return the index of the first occurrence of n in numbers,
-    or None if n doesn't appear in the list.
-    """
-    i = 0
-    for number in numbers:
-        if number == n:
-            return i
-        i += 1
+def str_to_int(s: str) -> Optional[int]:
+    """Return the integer representation of the string s, or None if it cannot be converted."""
+    try:
+        return int(s)
+    except ValueError:
+        return  # Error: this should be `return None` instead.
 
 
-def day_name_to_number(day: str) -> int:
-    """Return a number between 0-6 representing the given day of the week."""
-    if day == 'Monday':
-        return 0
-    elif day == 'Tuesday':
-        return 1
-    elif day == 'Wednesday':
-        return 2
-    elif day == 'Thursday':
-        return 3
-    elif day == 'Friday':
-        return 4
-    elif day == 'Saturday':
-        return 5
-    elif day == 'Sunday':
-        return 6
+if __name__ == "__main__":
+    import python_ta
+    python_ta.check_all()
