+++ examples/custom_checkers/r9711_missing_return_statement.py
@@ -1,26 +1,35 @@
-def missing_return() -> int:
-    print("no return")
+from typing import List, Optional
 
 
-def missing_return_in_branch() -> int:
-    a = 1
-    if a > 3:
-        print("no return")
-    else:
-        return a
+def index_of(numbers: List[int], n: int) -> Optional[int]:
+    """Return the index of the first occurrence of n in numbers,
+    or None if n doesn't appear in the list.
+    """
+    i = 0
+    for number in numbers:
+        if number == n:
+            return i
+        i += 1
 
 
-def multiple_branches() -> int:
-    if False:
+def day_name_to_number(day: str) -> int:
+    """Return a number between 0-6 representing the given day of the week."""
+    if day == 'Monday':
+        return 0
+    elif day == 'Tuesday':
         return 1
-    elif True:
-        print("no return")
-    else:
+    elif day == 'Wednesday':
         return 2
+    elif day == 'Thursday':
+        return 3
+    elif day == 'Friday':
+        return 4
+    elif day == 'Saturday':
+        return 5
+    elif day == 'Sunday':
+        return 6
 
 
-def try_except() -> int:
-    try:
-        print("try block")
-    except Exception:
-        print("except block")
+if __name__ == "__main__":
+    import python_ta
+    python_ta.check_all()
