+++ python_ta/checkers/inconsistent_or_missing_returns_checker.py
@@ -15,19 +15,14 @@ class InconsistentReturnChecker(BaseChecker):
 
     name = "inconsistent-or-missing-returns"
     msgs = {
-        # as R1710 is being disabled, we replace it with an identical message
         "R9710": (
-            """This function has at least one case where the function body will execute without ending in an explicit return statement.
-            You should check your code to make sure every possible execution path through the function body ends in a return statement.
-            Note: one common source of this error is if you're using if statements without an explicit else branch.
-            In this case, you should consider revising your code to either add an else branch,
-            or, if you are confident that the if and elif conditions cover all possible cases,
-            "you can convert the final "elif " into an " else ".",""",
+            """This function has inconsistent return statements: it sometimes returns a non-None value and sometimes uses `return` without a value.
+            To ensure clarity and consistency, every return statement should either return a value or explicitly return None""",
             "inconsistent-returns",
-            "Used to replace R1710 message inconsistent-return-statements",
+            "Used when a function has non-None return values at some cases and `return` in other cases",
         ),
         "R9711": (
-            "Missing return statement in function",
+            "Missing return statement in function `%s`",
             "missing-return-statement",
             "Used when a function does not have a return statement and whose return type is not None",
         ),
@@ -42,14 +37,15 @@ class InconsistentReturnChecker(BaseChecker):
 
     def _check_return_statements(self, node) -> None:
         """
-        Construct a CFG from the function. Check for inconsistent returns if there are
-        multiple return statements, and missing return statements if there are none.
+        Check for inconsistent returns and missing returns based on CFG
+        if the function has a return type annotation or has at least one
+        non-None return values.
         """
         if isinstance(node.returns, nodes.Const) and node.returns.value is None:
             return
 
         has_return_annotation = node.returns is not None
-        any_branches_have_return = False
+        has_return_value = False
 
         # get the blocks connected to the end of cfg
         end = node.cfg.end
@@ -63,16 +59,19 @@ class InconsistentReturnChecker(BaseChecker):
             if isinstance(statement, nodes.Return):
                 return_statements[block] = statement
                 if return_statements[block].value is not None:
-                    any_branches_have_return = True
+                    has_return_value = True
 
         # check for inconsistent or missing returns
-        if has_return_annotation or any_branches_have_return:
+        if has_return_annotation or has_return_value:
             for block in return_statements:
                 statement = return_statements[block]
                 if statement is None:
                     # for rendering purpose, the line is set to the last line of the function branch where return statement is missing
                     self.add_message(
-                        "missing-return-statement", node=node, line=block.statements[-1].fromlineno
+                        "missing-return-statement",
+                        node=node,
+                        line=block.statements[-1].fromlineno,
+                        args=node.name,
                     )
                 elif statement.value is None:
                     self.add_message("inconsistent-returns", node=statement)
